+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- cupti: add activity.cu;
 - cuda: add pti subdir;
 - python: add pip subdir;
 - shell/commands: add arp and ethtool subdir;